+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Chg: Have a proper hierarchy of Ultimate Notion exception classes.
 - New: The schema of a database provides a `to_pydantic_model` class for evaluating input.
 - New: More schema validations like checking for a title property and distinct property names.
+- Fix: Fix error when setting a (multi-)select option with no color specified.
 
 ## Version 0.7.1, 2025-01-05